refactor: task.py moved to top level contract.py
